requirements.txt
blis==0.2.4
certifi==2021.5.30
chardet==4.0.0
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==8.0.1
colorama==0.4.4
cymem==2.0.5
et-xmlfile==1.1.0
Flask==2.0.1
h5py==2.10.0
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==2.6.0
MarkupSafe==2.0.1
mathparse==0.1.2
murmurhash==1.0.5
numpy==1.21.0
openpyxl==3.0.7
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==3.13
requests==2.25.1
scipy==1.4.1
six==1.16.0
spacy==2.1.0
SQLAlchemy==1.3.24
srsly==1.0.5
thinc==7.0.8
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
nltk==3.6.2
Flask
flask_cors
nltk
pyspellchecker
 typing